romio: silence warnings in debug prints

Approved-by: Rob Latham <robl@mcs.anl.gov>

+++ src/mpi/romio/adio/common/ad_read_coll.c
@@ -339,10 +339,10 @@ void ADIOI_Calc_my_off_len(ADIO_File fd, int bufcount, MPI_Datatype
 #ifdef RDCOLL_DEBUG
         {
             int ii;
-            DBG_FPRINTF(stderr, "flattened %3lld : ", flat_file->count);
+            DBG_FPRINTF(stderr, "flattened %3lld : ", (long long) flat_file->count);
             for (ii = 0; ii < flat_file->count; ii++) {
-                DBG_FPRINTF(stderr, "%16lld:%-16lld", flat_file->indices[ii],
-                            flat_file->blocklens[ii]);
+                DBG_FPRINTF(stderr, "%16lld:%-16lld", (long long) flat_file->indices[ii],
+                            (long long) flat_file->blocklens[ii]);
             }
             DBG_FPRINTF(stderr, "\n");
         }
